不要なコードを削除したりまとめたりした
