--- Entity+CoreDataClass.swift
@@ -1,15 +1,0 @@
-//
-//  Entity+CoreDataClass.swift
-//  ShoppingApp
-//
-//  Created by koala panda on 2022/09/29.
-//
-//
-
-//import Foundation
-//import CoreData
-//
-//@objc(Entity)
-//public class Entity: NSManagedObject {
-//
-//}
