+++ ShoppingApp/Model/DataModel.swift
@@ -16,17 +16,6 @@ FirebaseのCloud Firestoreの読み書きするコードを記載します。
 import Foundation
 import FirebaseFirestore
 
-
-struct ItemDataType: Identifiable {
-    public var id: String
-    public var title: String
-    public var label: Int16
-    public var favorite: Bool
-    public var checked: Bool
-    public var finished: Bool
-    public var timestamp: Date
-}
-
 class ItemViewModel: ObservableObject {
     @Published var itemList = [ItemDataType]()
     @Published var filterdList0 = [ItemDataType]()
